Polishing Particle

+++ src/Color.cpp
@@ -5,6 +5,6 @@
 #include <SDL.h>
 
 SDL_Color Color::ToSDL(){
-    SDL_Color color = {r, g, b, a};
+    SDL_Color color = {(Uint8)r, (Uint8)g, (Uint8)b, (Uint8)a};
     return color;
 }
